Refactor: adjust namespaces
