+++ Filc/Services/Interfaces/EntityBasedInterfaces/ILessonService.cs
@@ -1,6 +1,7 @@
-﻿namespace Filc.Services.New_intefaces.EntityBasedInterfaces
+﻿namespace Filc.Services.Interfaces.EntityBasedInterfaces
 {
     public interface ILessonService
     {
+
     }
 }
