+++ Filc/Services/Interfaces/EntityBasedInterfaces/IParentService.cs
@@ -1,4 +1,4 @@
-﻿namespace Filc.Services.New_intefaces.EntityBasedInterfaces
+﻿namespace Filc.Services.Interfaces.EntityBasedInterfaces
 {
     public interface IParentService
     {
