+++ Filc/Services/Interfaces/EntityBasedInterfaces/IStudentService.cs
@@ -1,4 +1,4 @@
-﻿namespace Filc.Services.New_intefaces
+﻿namespace Filc.Services.Interfaces.EntityBasedInterfaces
 {
     public class IStudentService
     {
