+++ Filc/Services/Interfaces/EntityBasedInterfaces/ITeacherService.cs
@@ -1,4 +1,4 @@
-﻿namespace Filc.Services.New_intefaces
+﻿namespace Filc.Services.Interfaces.EntityBasedInterfaces
 {
     public class ITeacherService
     {
